IsRunning fixed
